Fixed a bug where an enemy would come back to life
Enemies are "destroyed" by being moved out of the screen.  This causes them
to be removed eventually in the update loop.  However, they were still too
close the the border, meaning that they could still be attracted back to
the player.  End this non-sense by moving the enemies farther away.